Criação de model de deposito e validação de nome completo

+++ src/main/java/br/com/contas/api/domain/repository/ContaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.contas.api.domain.model.Conta;
 
-public interface ContaRepository extends JpaRepository<Conta, Integer>{
+public interface ContaRepository extends JpaRepository<Conta, Long>{
 
 }
